pyproject.toml
[tool.poetry]
name = "h-crypto-price-bot"
version = "1.1.2"
description = ""
authors = ["halon176 <acc.github@halon.cc>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-telegram-bot = "^21.5"
web3 = "^7.2.0"
feedparser = "^6.0.10"
pandas = "^2.2.2"
kaleido = "0.2.1"
plotly = "^5.22.0"
aiohttp = { extras = ["speedups"], version = "^3.9.5" }
pydantic-settings = "^2.6.1"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
telegramify-markdown = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["build", "dist", "venv", ".venv", ".git", ".tox", ".mypy_cache", ".pytest_cache", ".vscode", ".idea", "poetry.lock"]
line-length = 120
indent-width = 4

[lint.select]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.pytest.ini_options]
addopts = "-v"